+++ pkg/RelpConnection/RelpConnection.go
@@ -6,11 +6,11 @@ import (
 	"errors"
 	"github.com/teragrep/rlp_05/internal/Errors"
 	"github.com/teragrep/rlp_05/internal/RelpCommand"
+	"github.com/teragrep/rlp_05/internal/RelpFrame"
 	"github.com/teragrep/rlp_05/internal/RelpParser"
 	"github.com/teragrep/rlp_05/internal/RelpWindow"
 	"github.com/teragrep/rlp_05/pkg/RelpBatch"
 	"github.com/teragrep/rlp_05/pkg/RelpDialer"
-	"github.com/teragrep/rlp_05/pkg/RelpFrame"
 	"io"
 	"log"
 	"os"
